proper cast for cos(x)

+++ game/zweidee/galaga.cpp
@@ -112,8 +112,8 @@ int galaga::CGalaga::draw_starfield_vert(unsigned char * data) // cheap trick ..
     r = ((i + 1) % 3) * 255;
     g = ( i      % 3) * 255;
     b = ((i - 1) % 3) * 255;
-    fbuf2d.setPixel(data, cos(i) *                      i, (iloopy + i) % 64, r, g, b);
-    fbuf2d.setPixel(data, cos(i) * (fbuf2d.width - 2) - i, (iloopy + i) % 64, r, g, b);
+    fbuf2d.setPixel(data, cos((float)i) *                      i, (iloopy + i) % 64, r, g, b);
+    fbuf2d.setPixel(data, cos((float)i) * (fbuf2d.width - 2) - i, (iloopy + i) % 64, r, g, b);
     // funny effect:
 //    fbuf2d.setPixel(data, (unsigned int)cos(i) *                      i, (iloopy + i) % 64, r, g, b);
 //    fbuf2d.setPixel(data, (unsigned int)cos(i) * (fbuf2D.width - 2) - i, (iloopy + i) % 64, r, g, b);
